+++ src/App.variables.scss
@@ -1,17 +1,3 @@
-// $violet: #484283;
-// $black: #444B58;
-// $white: #FFFFFF;
-// $red: #F14F4F;
-// $dark-red: #8C2828;
-// $light-red: #F8AFAF;
-// $light-gray1:#F6F6F6;
-// $light-gray2:#D5D5D5;
-// $light-gray3:#C4C4C4;
-// $gray: #B2B5BB;
-// $dark-gray:#646464;
-// $hover-gray: #87878740;
-
-
 $bg-dark: #484283;
 $text-white: rgba(255, 255, 255, 0.12);
 $link-gray: #c4c4c4;
