+++ src/assets/fonts/Jost-Bold.woff2
