+++ src/components/Accordion/Accordion.module.scss
@@ -1,9 +1,6 @@
 @import '../../App.variables.scss';
 
 .accordion {
-  font-size: 20px;
-  line-height: 30px;
-  text-align: left;
   padding: 25px 0;
   background: $bg-dark;
   color: $white;
@@ -13,6 +10,11 @@
   display: flex;
   align-items: center;
   justify-content: space-between;
+
+  h3 {
+    font-size: 20px;
+    line-height: 30px;
+  }
 
   & img.active {
     transform: rotate(45deg);
