+++ src/index.scss
@@ -2,41 +2,41 @@
 
 @font-face {
   font-family: "Jost";
-  src: url("Jost-Regular.woff2") format("woff2"),
-       url("Jost-Regular.woff") format("woff"),
-       url("Jost-Regular.ttf") format("ttf");
+  src: url("./assets/fonts/Jost-Regular.woff2") format("woff2"),
+       url("./assets/fonts/Jost-Regular.woff") format("woff"),
+       url("./assets/fonts/Jost-Regular.ttf") format("ttf");
   font-weight: 400;
 }
 
 @font-face {
   font-family: "Jost";
-  src: url("Jost-Medium.woff2") format("woff2"),
-       url("Jost-Medium.woff") format("woff"),
-       url("Jost-Medium.ttf") format("ttf");
+  src: url("./assets/fonts/Jost-Medium.woff2") format("woff2"),
+       url("./assets/fonts/Jost-Medium.woff") format("woff"),
+       url("./assets/fonts/Jost-Medium.ttf") format("ttf");
   font-weight: 500;
 }
 
 @font-face {
   font-family: "Jost";
-  src: url("Jost-SemiBold.woff2") format("woff2"),
-       url("Jost-SemiBold.woff") format("woff"),
-       url("Jost-SemiBold.ttf") format("ttf");
+  src: url("./assets/fonts/Jost-SemiBold.woff2") format("woff2"),
+       url("./assets/fonts/Jost-SemiBold.woff") format("woff"),
+       url("./assets/fonts/Jost-SemiBold.ttf") format("ttf");
   font-weight: 600;
 }
 
 @font-face {
   font-family: "Jost";
-  src: url("Jost-Bold.woff2") format("woff2"),
-       url("Jost-Bold.woff") format("woff"),
-       url("Jost-Bold.ttf") format("ttf");
+  src: url("./assets/fonts/Jost-Bold.woff2") format("woff2"),
+       url("./assets/fonts/Jost-Bold.woff") format("woff"),
+       url("./assets/fonts/Jost-Bold.ttf") format("ttf");
   font-weight: 700;
 }
 
 @font-face {
   font-family: "Jost";
-  src: url("Jost-ExtraBold.woff2") format("woff2"),
-       url("Jost-ExtraBold.woff") format("woff"),
-       url("Jost-ExtraBold.ttf") format("ttf");
+  src: url("./assets/fonts/Jost-ExtraBold.woff2") format("woff2"),
+       url("./assets/fonts/Jost-ExtraBold.woff") format("woff"),
+       url("./assets/fonts/Jost-ExtraBold.ttf") format("ttf");
   font-weight: 800;
 }
 
@@ -81,6 +81,15 @@ button {
 }
 img {
   max-width: 100%;
+}
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  font-size: inherit;
+  font-weight: inherit;
 }
 
 /* Общие стили */
@@ -139,6 +148,8 @@ body {
 }
 
 .add-button {
+  display: flex;
+  align-items: center;
   width: 50px;
   height: 50px;
   padding: 16px;
@@ -158,6 +169,7 @@ body {
   color: $white;
   transition: .3s;
   cursor: pointer;
+
   &:hover, &:active {
     color:  $link-gray;
   }
