+++ src/pages/HomePage/HomePage.module.scss
@@ -53,6 +53,11 @@
   }
 }
 
+.catalogBtn {
+  display: block;
+  margin: 0 auto;
+}
+
 .faqSection {
   padding: 60px 0;
 }
@@ -63,44 +68,4 @@
 }
 .faqTitle {
   margin-bottom: 20px;
-}
-
-.accordion {
-  font-size: 20px;
-  line-height: 30px;
-  text-align: left;
-  padding: 25px 0;
-  background: $bg-dark;
-  color: $white;
-  width: 100%;
-  border-top: 1px solid $white;
-  //transition: .4s;
-  cursor: pointer;
-  position: relative;
-  &::after {
-    content: '';
-    position: absolute;
-    top: 28px;
-    right: 0;
-    background: url("../../assets/add-icon.svg") center no-repeat;
-    width: 25px;
-    height: 25px;
-    .active {
-      transform: rotate(45deg);
-    }
-  }
-}
-
-.accordionText {
-  cursor: pointer;
-  width: 100%;
-  font-size: 16px;
-  line-height: 26px;
-  font-weight: 600;
-  max-height: 0;
-  overflow: hidden;
-  transition: max-height 0.2s ease-out;
-  p {
-    padding-bottom: 20px;
-  }
 }
