+++ src/pages/ProductDetail/ProductDetail.module.scss
@@ -13,18 +13,19 @@
   font-weight: 400;
 }
 .imgWrapper {
-  width: 520px;
+  flex: 0 0 520px;
 }
 .thumbnailWrapper {
   display: flex;
   flex-wrap: wrap;
   gap: 20px;
   margin-top: 10px;
+  img {
+    width: 70px;
+    height: 70px;
+  }
 }
-.thumbnailImg {
-  width: 70px;
-  height: 70px;
-}
+
 .productTitle {
   margin-bottom: 16px;
 }
@@ -45,7 +46,7 @@
   line-height: 24px;
   color: $red;
   font-weight: 500;
-  margin-bottom: 36px;
+  margin-bottom: 24px;
   padding: 12px 0;
   border-top: 1px solid $link-gray;
   border-bottom: 1px solid $link-gray;
@@ -65,6 +66,7 @@
 }
 .buy {
   display: flex;
+  flex-wrap: wrap;
   align-items: center;
   justify-content: space-between;
   padding: 12px;
